feat: 디자인 시스템 테이블 컴포넌트 적용
