+++ apps/admin/app/students/_components/StudentList.tsx
@@ -1,9 +1,9 @@
-import { css } from "@styled-system/css";
-import { styled } from "@styled-system/jsx";
 import { Text } from "@wow-class/ui";
+import Link from "next/link";
+import type { CSSProperties } from "react";
 import type { StudyStudentApiResponseDto } from "types/dtos/studyStudent";
-
-import StudentListItem from "./StudentListItem";
+import Table from "wowds-ui/Table";
+import TextButton from "wowds-ui/TextButton";
 
 const StudentList = ({
   studentList,
@@ -13,38 +13,43 @@ const StudentList = ({
   if (!studentList.length) return <Text>스터디 수강생이 없어요.</Text>;
 
   return (
-    <styled.table borderCollapse="collapse">
-      <thead>
-        <styled.tr borderBottom="1px solid" borderColor="sub">
-          <Text as="th" className={tableThStyle} typo="body2">
-            이름
-          </Text>
-          <Text as="th" className={tableThStyle} typo="body2">
-            학번
-          </Text>
-          <Text as="th" className={tableThStyle} typo="body2">
-            디스코드 사용자명
-          </Text>
-          <Text as="th" className={tableThStyle} typo="body2">
-            디스코드 닉네임
-          </Text>
-          <Text as="th" className={tableThStyle} typo="body2">
-            깃허브 링크
-          </Text>
-        </styled.tr>
-      </thead>
-      <tbody>
-        {studentList.map((student) => (
-          <StudentListItem key={student.memberId} {...student} />
-        ))}
-      </tbody>
-    </styled.table>
+    <Table>
+      <Table.Thead>
+        <Table.Th>이름</Table.Th>
+        <Table.Th>학번</Table.Th>
+        <Table.Th>디스코드 사용자명</Table.Th>
+        <Table.Th>디스코드 닉네임</Table.Th>
+        <Table.Th>깃허브 링크</Table.Th>
+      </Table.Thead>
+      <Table.Tbody>
+        {studentList.map((student) => {
+          const { name, studentId, discordUsername, nickname, githubLink } =
+            student;
+          return (
+            <Table.Tr key={studentId} value={studentId}>
+              <Table.Td>{name}</Table.Td>
+              <Table.Td>{studentId}</Table.Td>
+              <Table.Td>{discordUsername}</Table.Td>
+              <Table.Td>{nickname}</Table.Td>
+              <Table.Td>
+                <TextButton
+                  asProp={Link}
+                  href={githubLink || ""}
+                  style={textButtonStyle}
+                  text={githubLink}
+                />
+              </Table.Td>
+            </Table.Tr>
+          );
+        })}
+      </Table.Tbody>
+    </Table>
   );
 };
 
-const tableThStyle = css({
-  padding: "1rem",
-  textAlign: "left",
-});
+const textButtonStyle: CSSProperties = {
+  width: "fit-content",
+  padding: 0,
+};
 
 export default StudentList;
